refactor solver selection

+++ src/main.rs
@@ -6,7 +6,7 @@ fn main() {
     let day = std::env::args()
         .nth(1)
         .expect("Day should be passed as first argument")
-        .parse::<u8>()
+        .parse::<usize>()
         .expect("Cannot parse day as number");
     let input_bytes = fs::read(format!("./input/day{day}.txt"))
         .unwrap_or_else(|_| panic!("No input for day {day}"));
